[CHORE] model 파일 분리 (#100)
