+++ Maddori.Apple/Maddori.Apple/Network/Reqeust/DTO/AddFeedbackDTO.swift
@@ -14,8 +14,8 @@ struct AddFeedbackDTO: Encodable {
     let start_content: String
     let to_id: Int
     
-    enum FeedBackDTO: Encodable {
-        case continueType
-        case stopType
+    enum FeedBackDTO: String, Encodable {
+        case continueType = "Continue"
+        case stopType = "Stop"
     }
 }
