+++ Maddori.Apple/Maddori.Apple/Network/Response/EditFeedBackResponse.swift
@@ -10,17 +10,3 @@ import Foundation
 struct EditFeedBackResponse: Decodable {
     let feedback: FeedBackResponse?
 }
-
-struct FeedBackResponse: Decodable {
-    let id: Int?
-    let type: String?
-    let keyword: String?
-    let content: String?
-    let start_content: String?
-    let from_id: Int?
-    let to_id: Int?
-    let team_id: Int?
-    let reflection_id: Int?
-    let reflection: ReflectionResponse?
-    let user: MemberResponse?
-}
